1. LIF series loading is now lazily loaded and thread safe. 2. Sequencer time marker is now connected to seek_position 3. Data point marker is changed to circle from diamond 4. Sequencer frame index is not int64_t 5. WIP layout mods 6. Added list option to command line: ./VisibleRunApp file.lif list will list all series in file 7. Added separate functions to setup loggers with gui or just text 8. WIP cleanup signal connections to Cinder

+++ Visible/include/imGuiCustom/ImSequencer.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <cstddef>
+#include <stdint.h>
 
 struct ImDrawList;
 struct ImRect;
@@ -20,8 +21,8 @@ namespace ImSequencer
     struct SequenceInterface
     {
         bool focused = false;
-        virtual int GetFrameMin() const = 0;
-        virtual int GetFrameMax() const = 0;
+        virtual int64_t GetFrameMin() const = 0;
+        virtual int64_t GetFrameMax() const = 0;
         virtual size_t GetItemCount() const = 0;
 
         virtual void BeginEdit(int /*index*/) {}
@@ -46,6 +47,6 @@ namespace ImSequencer
 
 
     // return true if selection is made
-    bool Sequencer(SequenceInterface *sequence, int *currentFrame, bool *expanded, int *selectedEntry, int *firstFrame, int sequenceOptions);
+    bool Sequencer(SequenceInterface *sequence, int64_t *currentFrame, bool *expanded, int *selectedEntry, int64_t *firstFrame, int sequenceOptions);
 
 }
